feat: implement sign in with linkedin
create auth, callback, and token cloud functions to assist
the 3-legged OAuth sign-in flow with linkedin.

auth is a callable cloud function that creates a firestore object
with origin and redirect properties.
the id of this created firestore object is the state for the sign-in flow.
if this is the first time the LinkedIn user is signing in, this state
id string will end up being the uid of the user in Firebase Auth,
also this object would now hold the profile data of the user.

callback is an onRequest https cloud function that is called by linkedin.
it will have state and code attached to its query parameters by linkedin.
callback verifies that a firestore document with id of state exists.
it then obtains OAuth access_token from linkedin API through the code.
callback then gets user's profile info from various linkedin API endpoints.

if the LinkedIn user's id already exists in the database,
firebase auth user object and firestore user object are updated with
the new profile data. also, the state id string is changed to
that of this existing user's uid. the previously created state
firestore object is deleted.

if however the LinkedIn user is a new user, their profile data and the
state id string are used to create a new firebase auth user object.
their profile data is added to the firestore  state object and the origin
and redirect properties are deleted. the state string becomes the uid.

in both cases of new or existing users, a linkedin object is stored in an
auth subcollection of the user in firestore. this permits searching
if the user exists has signed in before on next sign in.
this linkedin object contains linkedin user id, signInTime, OAuth
access_token (from LinkedIn) and when it expires.

callback cloud function redirects to the origin of the first request
with the state string attached as a query in the url.

token is a callable cloud function that takes state string
(which is now technically a uid of a signed in user), and returns a
custom token for firebase auth to sign in the user in the frontend.
add matsnackbar for feedback in the frontend